refactor(session replay): move events storage logic in idb to separate class
